add an option to disable logging
